third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/workers/access-control-basic-get-fail-non-simple-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-get-fail-non-simple.cgi. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
GET should not trigger a preflight request from a worker unless it has non-simple headers.

PASS: Cross-domain access allowed for simple get.
PASS: Exception thrown. Cross-domain access was denied in 'send'. [A network error occurred.].
DONE
